+++ routes/__init__.py
@@ -3,6 +3,7 @@ from flask import Flask
 from routes.auth_routes import auth_bp
 from routes.dashboard_routes import dashboard_bp
 from routes.demo_routes import demo_bp
+from routes.timezone_routes import timezone_bp
 from routes.transaction_routes import transaction_bp
 
 
@@ -22,3 +23,4 @@ def register_blueprints(app: Flask):
     app.register_blueprint(transaction_bp, url_prefix="/transactions")
     # app.register_blueprint(account_bp, url_prefix="/accounts")
     app.register_blueprint(demo_bp, url_prefix="/demo")
+    app.register_blueprint(timezone_bp, url_prefix="/timezone")
